common CustomURLValidation
